Add aws actuator
This is a port of https://github.com/StackPointCloud/kube-deploy/tree/aws-machineactuator/cluster-api-aws
as a sample of what adding a new provider to the ext-apiserver
might look like.

This commit only adds the provider, and implements (an untested)
actuator. The next step is to create a deployer that bootstraps
the cluster (which won't be done).